Time: 943 ms (37.02%), Space: 14.5 MB (54.38%) - LeetHub

+++ 873-length-of-longest-fibonacci-subsequence/873-length-of-longest-fibonacci-subsequence.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int n;
     unordered_map<int, int> idx;
-    int dp[1010][1010];
+    short dp[1010][1010];
     
     int lenLongestFibSubseq(vector<int>& arr) {
         n = arr.size();
